active BoardCalculation moved to own Saga
